fix: adjust user videos res

+++ src/videos/services/user-videos.service.ts
@@ -1,6 +1,6 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
-import { CreateVideoDto } from '../dto/create-video.dto';
-import { UpdateVideoDto } from '../dto/update-video.dto';
+import { UserCreateVideoDto } from '../dto/create-video.dto';
+import { UserUpdateVideoDto } from '../dto/update-video.dto';
 import { Video, VideoDocument } from '../entities/video.entity';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
@@ -13,45 +13,57 @@ export class UserVideosService {
     private videoModel: Model<VideoDocument>,
   ) {}
 
-  public async create(dto: CreateVideoDto): Promise<VideoResponseDto> {
-    const created = await this.videoModel.create(dto);
+  public async createToUser(
+    user_id: string,
+    dto: UserCreateVideoDto,
+  ): Promise<VideoResponseDto> {
+    const created = await this.videoModel.create({ ...dto, user_id });
 
     return new VideoResponseDto(created);
   }
 
-  public async findAll(): Promise<VideoResponseDto[]> {
-    const videos = await this.videoModel.find();
+  public async findAllVideosOfUser(
+    user_id: string,
+  ): Promise<VideoResponseDto[]> {
+    const videos = await this.videoModel.find({ user_id });
+
     return videos.map((video) => new VideoResponseDto(video));
   }
 
-  private async findVideoByID(_id: string): Promise<Video> {
-    const video = await this.videoModel.findById(_id);
+  private async findVideoByIDOfUser(
+    _id: string,
+    user_id: string,
+  ): Promise<Video> {
+    const video = await this.videoModel.findOne({ _id, user_id });
 
-    if (!video) throw new NotFoundException('Video not found');
+    if (!video) throw new NotFoundException('Video não encontrada');
 
     return video;
   }
 
-  public async findOne(_id: string): Promise<VideoResponseDto> {
-    const video = await this.findVideoByID(_id);
-    return new VideoResponseDto(video);
-  }
-
-  public async update(
+  public async findOneVideoOfUser(
     _id: string,
-    dto: UpdateVideoDto,
+    user_id: string,
   ): Promise<VideoResponseDto> {
-    await this.findVideoByID(_id);
+    const category = await this.findVideoByIDOfUser(_id, user_id);
 
-    const rawData = { ...dto };
-
-    await this.videoModel.updateOne({ _id }, rawData);
-
-    return this.findOne(_id);
+    return new VideoResponseDto(category);
   }
 
-  public async remove(_id: string): Promise<void> {
-    await this.findVideoByID(_id);
-    await this.videoModel.deleteOne({ _id });
+  public async updateVideoOfUser(
+    _id: string,
+    user_id: string,
+    dto: UserUpdateVideoDto,
+  ): Promise<VideoResponseDto> {
+    await this.findVideoByIDOfUser(_id, user_id);
+
+    await this.videoModel.updateOne({ _id, user_id }, dto);
+
+    return new VideoResponseDto(await this.findVideoByIDOfUser(_id, user_id));
+  }
+
+  public async remove(_id: string, user_id: string): Promise<void> {
+    await this.findVideoByIDOfUser(_id, user_id);
+    await this.videoModel.deleteOne({ _id, user_id });
   }
 }
